+++ src/main/java/com/example/equipmentmanagement/service/EquipmentTypeService.java
@@ -2,9 +2,11 @@ package com.example.equipmentmanagement.service;
 
 import com.example.equipmentmanagement.dto.equipment.EquipmentTypeDto;
 import com.example.equipmentmanagement.dto.equipment.EquipmentTypeSaveDto;
+import com.example.equipmentmanagement.exception.BadRequestAlertException;
 import com.example.equipmentmanagement.mapper.EquipmentTypeMapper;
 import com.example.equipmentmanagement.exception.ResourceNotFoundException;
 import com.example.equipmentmanagement.model.EquipmentType;
+import com.example.equipmentmanagement.repository.EquipmentRepository;
 import com.example.equipmentmanagement.repository.EquipmentTypeRepository;
 import org.springframework.stereotype.Service;
 
@@ -17,9 +19,11 @@ import static com.example.equipmentmanagement.mapper.EquipmentTypeMapper.toEntit
 public class EquipmentTypeService {
 
     private final EquipmentTypeRepository equipmentTypeRepository;
+    private final EquipmentRepository equipmentRepository;
 
-    public EquipmentTypeService(EquipmentTypeRepository equipmentTypeRepository) {
+    public EquipmentTypeService(EquipmentTypeRepository equipmentTypeRepository, EquipmentRepository equipmentRepository) {
         this.equipmentTypeRepository = equipmentTypeRepository;
+        this.equipmentRepository = equipmentRepository;
     }
 
     public List<EquipmentTypeDto> getAllEquipmentTypes() {
@@ -47,6 +51,11 @@ public class EquipmentTypeService {
 
     public void deleteEquipmentType(Long id) {
         EquipmentType existingEquipmentType = findEquipmentTypeById(id);
+        int count = equipmentRepository.countByTypeId(id);
+        if (count > 0) {
+            throw new BadRequestAlertException("Failed. There are equipment with this type assigned: " + count);
+        }
+
         this.equipmentTypeRepository.delete(existingEquipmentType);
     }
